+++ user/src/main/java/top/jhechem/user/service/impl/FunctionAuthServiceImpl.java
@@ -26,7 +26,7 @@ public class FunctionAuthServiceImpl implements FunctionAuthService {
         return auths;
     }
 
-    public Set<Integer> getRanges(String path, long adminId) {
+    public Set<Integer> getRanges(String path, int adminId) {
         return mapper.getRanges(path, adminId);
     }
 }
